refactor(core): remove props default getter
Default getter returns the last cached prop which causes the virtual dom differ to not set the new property thus skips the prop event. For the system to be successfully reactive change should be detected at component level only and if needed optimizations should be using immutability.

BREAKING CHANGE: props should be accessed via `.props[<prop name>]` and not directly